Moving xlgui.icons.IconManager.pixbuf_from_data to xlgui.guiutil.pixbuf_from_data
Makes more sense
