+++ xlgui/guiutil.py
@@ -35,7 +35,6 @@ from gi.repository import Gtk
 from gi.repository import Pango
 
 from xl import settings, xdg
-from xlgui import icons
 
 # Import from external namespace
 from xl.externals.gi_composites import GtkTemplate as _GtkTemplate
@@ -151,6 +150,67 @@ def gtk_widget_replace(widget, replacement):
     return replacement
 
 
+def pixbuf_from_data(data, size=None, keep_ratio=True, upscale=False):
+    """
+        Generates a pixbuf from arbitrary image data
+
+        :param data: The raw image data
+        :type data: byte
+        :param size: Size to scale to; if not specified,
+            the image will render to its native size
+        :type size: tuple of int
+        :param keep_ratio: Whether to keep the original
+            image ratio on resizing operations
+        :type keep_ratio: bool
+        :param upscale: Whether to upscale if the requested
+            size exceeds the native size
+        :type upscale: bool
+
+        :returns: the generated pixbuf
+        :rtype: :class:`GdkPixbuf.Pixbuf` or None
+    """
+    if not data:
+        return None
+
+    pixbuf = None
+    loader = GdkPixbuf.PixbufLoader()
+
+    if size is not None:
+        def on_size_prepared(loader, width, height):
+            """
+                Keeps the ratio if requested
+            """
+            if keep_ratio:
+                scale = min(size[0] / float(width), size[1] / float(height))
+
+                if scale > 1.0 and upscale:
+                    width = int(width * scale)
+                    height = int(height * scale)
+                elif scale <= 1.0:
+                    width = int(width * scale)
+                    height = int(height * scale)
+            else:
+                if upscale:
+                    width, height = size
+                else:
+                    width = height = max(width, height)
+
+            loader.set_size(width, height)
+        loader.connect('size-prepared', on_size_prepared)
+
+    try:
+        loader.write(data)
+        loader.close()
+    except GLib.GError as e:
+        logger.warning('Failed to get pixbuf from data: {error}'.format(
+            error=e.message
+        ))
+    else:
+        pixbuf = loader.get_pixbuf()
+
+    return pixbuf
+
+
 class ScalableImageWidget(Gtk.Image):
     """
         Custom resizeable image widget
@@ -192,7 +252,7 @@ class ScalableImageWidget(Gtk.Image):
         """
         if not data:
             return
-        pixbuf = icons.MANAGER.pixbuf_from_data(data)
+        pixbuf = pixbuf_from_data(data)
         self.set_image_pixbuf(pixbuf, fill)
 
     def set_image_pixbuf(self, pixbuf, fill=False):
